move SdkVersion and fix tests
