+++ relay-event-normalization/src/event.rs
@@ -6,7 +6,6 @@ use std::collections::hash_map::DefaultHasher;
 
 use std::hash::{Hash, Hasher};
 use std::mem;
-use std::num::ParseIntError;
 use std::sync::OnceLock;
 
 use itertools::Itertools;
@@ -28,6 +27,7 @@ use smallvec::SmallVec;
 use uuid::Uuid;
 
 use crate::normalize::request;
+use crate::sdk_version::SdkVersion;
 use crate::span::ai::normalize_ai_measurements;
 use crate::span::tag_extraction::extract_span_tags_from_event;
 use crate::utils::{self, get_event_user_tag, MAX_DURATION_MOBILE_MS};
@@ -1480,72 +1480,6 @@ fn normalize_app_start_measurements(measurements: &mut Measurements) {
     }
     if let Some(app_start_warm_value) = measurements.remove("app.start.warm") {
         measurements.insert("app_start_warm".to_string(), app_start_warm_value);
-    }
-}
-
-/// Represents a SDK version using the format "<major>.<minor>.<patch>".
-#[derive(Debug, PartialOrd, Eq, PartialEq, Ord)]
-struct SdkVersion {
-    major: usize,
-    minor: usize,
-    patch: usize,
-    release_type: ReleaseType,
-}
-
-/// Represents the release type which might be present in a version string,
-/// for example: 9.0.0-alpha.0.
-/// Release types are also comparable to each other, using the rules:
-/// alpha < beta < release.
-/// **NOTE**: The discriminants are explicitly set to keep the comparison rules
-///           even if the order is changed.
-#[derive(Debug, Eq, PartialEq, Ord, PartialOrd)]
-enum ReleaseType {
-    Alpha = 1,
-    Beta = 2,
-    Release = 3,
-}
-
-impl SdkVersion {
-    pub const fn new(major: usize, minor: usize, patch: usize) -> Self {
-        Self {
-            major,
-            minor,
-            patch,
-            release_type: ReleaseType::Release,
-        }
-    }
-
-    /// Attempts to parse a version string and returning a [`ParseIntError`] if it contains any
-    /// non-numerical characters apart from dots.
-    /// If a version part is not provided, 0 will be assumed.
-    /// For example:
-    /// 1.2 -> 1.2.0
-    /// 1   -> 1.0.0
-    pub fn from_str(input: &str) -> Result<Self, ParseIntError> {
-        let mut split = input.split(".");
-        let major = split.next().map(str::parse).transpose()?.unwrap_or(0);
-        let minor = split.next().map(str::parse).transpose()?.unwrap_or(0);
-        let (patch, release_type) = if let Some(next) = split.next() {
-            match next.split_once("-") {
-                Some((patch, release_type)) => (
-                    patch.parse()?,
-                    if release_type.starts_with("alpha") {
-                        ReleaseType::Alpha
-                    } else {
-                        ReleaseType::Beta
-                    },
-                ),
-                None => (next.parse()?, ReleaseType::Release),
-            }
-        } else {
-            (0, ReleaseType::Release)
-        };
-        Ok(Self {
-            major,
-            minor,
-            patch,
-            release_type,
-        })
     }
 }
 
@@ -4559,43 +4493,5 @@ mod tests {
           },
         }
         "#);
-    }
-
-    #[test]
-    pub fn test_version_compare() {
-        let main_version = SdkVersion::new(1, 2, 3);
-        let less = SdkVersion::new(1, 2, 1);
-        let greater = SdkVersion::new(2, 1, 1);
-        let equal = SdkVersion::new(1, 2, 3);
-        assert!(main_version > less);
-        assert!(main_version < greater);
-        assert_eq!(main_version, equal);
-    }
-
-    #[test]
-    pub fn test_version_string_compare() {
-        let main_version = SdkVersion::from_str("1.2.3").unwrap();
-        let less = SdkVersion::from_str("1.2.1").unwrap();
-        let greater = SdkVersion::from_str("2.1.1").unwrap();
-        let equal = SdkVersion::from_str("1.2.3").unwrap();
-        assert!(main_version > less);
-        assert!(main_version < greater);
-        assert_eq!(main_version, equal);
-    }
-
-    #[test]
-    pub fn test_release_type() {
-        let alpha = SdkVersion::from_str("9.0.0-alpha.2").unwrap();
-        let beta = SdkVersion::from_str("9.0.0-beta.2").unwrap();
-        let release = SdkVersion::from_str("9.0.0").unwrap();
-
-        assert!(alpha < beta);
-        assert!(beta < release);
-        assert!(alpha < release)
-    }
-
-    #[test]
-    pub fn test_version_string_parse_failed() {
-        assert!(SdkVersion::from_str("amd64").is_err());
     }
 }
