+++ relay-event-normalization/src/lib.rs
@@ -26,6 +26,8 @@ mod validation;
 
 pub use validation::{validate_event, validate_span, EventValidationConfig};
 pub mod replay;
+mod sdk_version;
+
 pub use event::{
     normalize_event, normalize_measurements, normalize_performance_score, NormalizationConfig,
 };
