+++ tests/integration/test_normalization.py
@@ -426,6 +426,38 @@ def test_empty_ip_not_auto(mini_sentry, relay, platform, version, expected):
 
 
 @pytest.mark.parametrize(
+    "platform, version, expected",
+    [
+        ("javascript", "8.55", "89.128.74.91"),
+        ("javascript", "9", "89.128.74.91"),
+        ("cocoa", "6.1.4", "89.128.74.91"),
+        ("cocoa", "6.2", "89.128.74.91"),
+        ("objc", "6.1.4", "89.128.74.91"),
+        ("objc", "6.2", "89.128.74.91"),
+        ("random_sdk", None, "89.128.74.91"),
+    ],
+)
+def test_explicit_ip_version_cutoff(mini_sentry, relay, platform, version, expected):
+    project_id = 42
+    relay = relay(mini_sentry)
+
+    mini_sentry.add_basic_project_config(project_id)
+
+    relay.send_event(
+        project_id,
+        {
+            "platform": platform,
+            "sdk": {"version": version},
+            "user": {"ip_address": "89.128.74.91"},
+        },
+    )
+
+    envelope = mini_sentry.captured_events.get(timeout=1)
+    event = envelope.get_event()
+    assert event["user"]["ip_address"] == expected
+
+
+@pytest.mark.parametrize(
     "scrub_ip_addresses, expected", [(True, None), (False, "127.0.0.1")]
 )
 def test_ip_address_scrubbed(mini_sentry, relay, scrub_ip_addresses, expected):
